Refactored common code into BaseFormController from BaseCifController, BaseIrfController and BaseVdfController. Fixed CIF issue with person identifiers when copying a person from IRF.
