+++ src/app/cif/india/step-templates/mainPv.html
@@ -6,7 +6,8 @@
                 <input type="text" class="form-control" placeholder="Name" ng-model="$ctrl.questions[292].response.name.value">
             </div>
             <div class="col-md-1">
-	            <associated-person ng-model="$ctrl.questions[292].response" person-list="$ctrl.associatedPersons" birth-date="$ctrl.dateData.questions[292]"></associatedPerson>
+	            <associated-person ng-model="$ctrl.questions[292].response" person-list="$ctrl.associatedPersons" birth-date="$ctrl.dateData.questions[292]"
+	            		id-choice="$ctrl.personIdentifierChoice.questions[292].radioValue"></associatedPerson>
             </div>
         </div>
         <div class="col-md-3 form-group">
