+++ src/app/cif/india/step-templates/personBoxes/personBoxModal.html
@@ -42,7 +42,9 @@
                     </div>
                 </div>
                 <div class="col-md-1 form-group">
-            		<associated-person ng-model="PersonBoxModalController.questions[9].response" person-list="PersonBoxModalController.associatedPersons" birth-date="PersonBoxModalController.dateData.questions[9]"></associatedPerson>
+            		<associated-person ng-model="PersonBoxModalController.questions[9].response" person-list="PersonBoxModalController.associatedPersons"
+            				birth-date="PersonBoxModalController.dateData.questions[9]"
+            				id-choice="PersonBoxModalController.personIdentifierChoice.questions[9].radioValue"></associatedPerson>
            		</div>
             </div>
             <div class="row">
