+++ src/app/irf/africa/africa.component.js
@@ -29,8 +29,8 @@ export class IrfAfricaController extends BaseIrfController {
         ];
     }
     
-    getIdentificationType() {
-        return "Passport";
+    getDefaultIdentificationTypes() {
+        return ['Passport'];
     }
 
     openIntercepteeModal(card, isAdd = false, idx = null) {
